feat: add existing content from `plco-analysis`, after a fashion
so the initial goal here was just to migrate in the code from `plco-analysis`. however, with the
primary code wrapped in a function, the code had some ridiculous cyclomatic complexity that caused
linting to fail. so I've spent all day refactoring, and now everything is 15 or lower as lintr
insists. this is a wrapped version of the code from `plco-analysis`; the code there will need to be
converted simply to a function call via a call to Rscript. there is enough change here that
extensive testing to be sure nothing was broken is required, in addition to the testing that was
already warranted.

re #1 #2