test: add end-to-end model matrix construction test for continuous trait
this patch adds testthat tests for the consensus overall result of the package for an exemplar
continuous trait. the test only runs if it can find the old plco-analysis installation; once this
thing goes faster, I'll add in a regression test style test based on the current run output. note
that for continuous traits (and transformed covariates), the inverse normal transform is using
*random* resolution of ties for better statistical properties. this has many benefits but means the
results are not exactly identical across multiple runs, even when things are working correctly.
therefore most of the model matrix is tested for identity but the phenotype itself is tested for
near-equality with a very high (currently 0.0075) precision tolerance

re #3