perf: switch `base::read.table` to 'data.table::fread` for phenotype database load
this package's major bottleneck has traditionally been the load operation on the backend phenotype
file. for three `testthat` end-to-end tests, the entire process took about two minutes, which is
just ridiculous. switching to `data.table::fread` has reduced the testing time to 30-40s, depending
on the attempt, which is just fabulous.

re #6