test: stabilize random categorical test; add test to confirm issue
ok so this was a whole process. one of the categorical tests was infrequently and randomly failing.
in the end, the reason was that if the category level 10 was tied in count with a single digit
non-reference category, the order of the output dummy variables would be based on lexicographical
sorting from the true code, but the test was using numerical sort. this has been resolved, and a
test has been added to detect the fact that it's definitely using lexicographical sort to resolve
ties. this has no impact whatsoever on downstream regression results.

re #3