test: add test confirming filtering works with multiple data types
so i'm not really sure how to handle input data in the *.control-inclusion and *.control-exclusion
files, as the category levels can be arbitrary strings that `read.table` will downcast if possible.
but as it turns out, it totally doesn't matter! R `%in%` just handles it. and i know this now,
because i added tests to that effect lol

re #5